power: smb1360: Implement Software JEITA
Default hard JEITA limits for SMB1360 are not configurable as they
are one time OTP programmed. The soft thresholds are configurable
from software.

To allow configurability of the JEITA hard thresholds, implement
a software based JEITA configuration logic. It uses reconfigurability
of the soft thresholds and the hot/cold soft notifications to achieve
this.

Note that the hard thresholds can only be re-configured if they lie
beyond the existing OTP configured values of [0, 55].

CRs-Fixed: 719537
Change-Id: I691f46ecd89abf5296ef5ca79251cea7513a0b54
Signed-off-by: Zhenhua Huang <zhenhuah@codeaurora.org>