msm: camera: work around for black screen issue
This change will provide a workaround for the black
screen issue during late camera. The camera thread
needs display pipes to show preview and guidance lane
and black screen was appearing due to pipe allocation
failure on mdp. The fix is that camera thread will wait
in a loop of 100 milliseconds till the pipe is released
by display HAL for camera preview and guidance lane.

Change-Id: Idf71958fabf136c220e31c11c81ba6ad8917c85f
Signed-off-by: Linto Paul <lpaul@codeaurora.org>