power: smb1360: Add a DT property to prevent system sleep under min_soc
When the soc(state of charge) becomes 0%, userspace should shut down the
device. In the current design, there is a possibility that the device
stays suspended even when the SOC reaches 0% leading to an under voltage
lock-out condition. Handle this by adding a DT node which prevents the
system from going to sleep if the SOC is under the min_soc threshold.

CRs-Fixed: 721515
Change-Id: I84bc46c5c9353261f72c996fbc25a745fa174738
Signed-off-by: Zhenhua Huang <zhenhuah@codeaurora.org>