usb: msm_otg: Add support for multiple phy instances
The USB PHY framework supports only 1 USB 2.0 PHY in 3.4 kernel. The
framework is extended to support multiple PHY in 3.10 kernel. But this
work is based on device tree model which is not supported for some
targets like APQ8064.

Add a MSM specific API called msm_otg_get_tranceiver() that can return
a PHY handle based on the platform id. The MSM OTG driver stores the
PHY handles in an array during the probe of msm_otg platform device.
The MSM UDC/HCD are required to call msm_get_transceiver() API instead
of the generic transceiver get API. Remaining all phy API can be called
once the phy handle is obtained.

Change-Id: Ice5a6e8af314c19806c6e0141d2992d35de6d63e
Signed-off-by: ChandanaKishori Chiluveru <cchilu@codeaurora.org>