hwspinlock/msm: correct pid for release functionality
Valid remote spinlock processor ID's are 1..N, but the processor IDs
start at 0.  To fix this, the value written to the HW Mutex is PID + 1.
Because the release functionality was using the PID for comparison, the
remote spinlock was not being properly released during subsystem restart
processing which would ultimately result in a watchdog if the remote
subsystem failed with the remote spinlock locked.

Update the release functions to compare the lock value with PID + 1.

CRs-Fixed: 725337
Change-Id: I4d484b5d37cfd1f9f8c6a3290ed378812620b130
Signed-off-by: Eric Holmberg <eholmber@codeaurora.org>
Signed-off-by: Arun Kumar Neelakantam <aneela@codeaurora.org>