ASoC: wcd9xxx: modify sequence post SSR
Perform mux configuration before MBHC init.
Otherwise, MLCK will not be enabled which results
in wrong initial calibration of dce and sta.
Updating the mux configuration for class h dsm mux
during codec bring up after ssr is causing the
widget not to be powered up as the dapm sees no
change in the register associated with it. The change
ensures the mux register is not updated post SSR.

Change-Id: I066c8e0f2d67eac22a080d99797acd8fb655d432
Signed-off-by: Aviral Gupta <aviralg@codeaurora.org>