ARM: dts: msm: add release dt property to synaptics node
devm_pinctrl_get() does not release the ownership of mux function
of pins in the associated pin-group whenever a failure occurs in
driver probe routine. i.e struct pin_desc's mux_owner field is
still marked as being in use after a failure occurs in probe.
As a result of this, if another driver tries to acquire the
ownership of same pin, it gets an error while applying that
setting. To fix this, explicitly release the mux function
ownership of the the pin, by adding a new pin-group in pinctrl
DT and a new pinctrl state in touch device's DT node. This new
pin-group does not have a function setting (qcom,pin-func property).
This new state is explicitly activated during a probe failure
and driver remove routine to release the mux function ownership.
Add release pinctrl dt property to synaptics v1 and
v2 driver for 8916 and 8939.

Change-Id: Id03c6886edc4362c04c0658e6cc8a4a548793235
Signed-off-by: Shantanu Jain <shjain@codeaurora.org>