net: usb: Add rmnet usb host driver snapshot
This snapshot is taken as of msm-3.10 commit:
 78c36fa0ef (Merge "msm: mdss: Prevent backlight update during
 continuous splash")
At the same time, fix compiler a handful of compiler warnings when
building 64-bit: use %zu for printing size_t, and unsigned int values
for cbits_tolocal and cbits_tomdm.

The rmnet usb host driver is used to communicate with modem devices
having an rmnet interface. This driver works as pass through layer
for control and data path communication.

Change-Id: I05a458c3a40e50f7a96219b8a3d8c1a06a0231c6
Signed-off-by: Matt Wagantall <mattw@codeaurora.org>