platform: msm: add Per-File-Tagger (PFT) driver snapshot
This snapshot is taken as of msm-3.10 commit:
 8e80dd5c0 (Merge "ASoC: wcd: Donot return IRQ_NONE")

While integrating, fix a number of compiler warnings related
to debug prints, and update the datatype used for various 'uid'
values to be of kuid_t type.

This driver is part of the Per-File-Encryption (PFE) feature.
It allows to tag enterprise files and encrypt them
while keeping the user private files untagged and plain text.

Change-Id: I2b832cc317e3a3253809eca2824e254b9fb2c043
Signed-off-by: Matt Wagantall <mattw@codeaurora.org>