usb: msm_otg: Handle spurious BSV clear interrupt
The DP manual pull-up flag is set for all the controllers to get
peripheral mode working without the VBUS routed to the PHY. The
controller is generating a spurious BSV clear upon cable disconnect
though the corresponding interrupt is disabled in the PHY and the
controller. The gadget is disconnected upon BSV clear but there will
not be a BSV set interrupt when the cable is connected back as the
otg mode is configured to USER_CONTROL. The peripheral mode stops
working after few cable connect and disconnect events. Fix this by
ignoring the spurious BSV clear interrupt.

Change-Id: Ice742f183c35234ac314753f554231f63b0c89b6
Signed-off-by: ChandanaKishori Chiluveru <cchilu@codeaurora.org>