usb_bam: Fix disconnect sequence for BAM2BAM_IPA mode
Due to confusing naming of functions in the disconnect
sequence, there was a decrement of the USB producer pipes
in the wrong place. Hence on USB disconnect case producer
pipe counter is decremented multiple times.
On next USB re-connect due to producer pipe counter not
being set to non-zero value, IPA resource manager API is not
called for requesting IPA producer resource which
causes a stall in USB tethered functionality. To fix this issue,
this change moves the decrement of USB producer pipe to API
where the USB producer pipe is actually released. Also add
proper comments to disconnect APIs to clear purpose of those
APIs.

Change-Id: I329788c61d4272579c0096c38cc1eeff94fa277c
Signed-off-by: Lena Salman <esalman@codeaurora.org>