Revert "arm64: PCI(e) arch support"
This reverts commit 32f20775e6400953915146d8888f096281f346f2.
This change adds the arch support for PCI(e) for arm64 on 3.10
kernel and is not compatible with the 3.14 kernel. The files
added or modfied in this patch conflict with the ones for PCI(e)
arm64 3.14 patch. Reverting this change resolves the conflicts
when applying the arch support for PCI(e) for arm64 patch for
3.14 kernel.

Change-Id: Ibe7f304c7969d25cf84ab233d135a568bedd77d7
Signed-off-by: Tony Truong <truong@codeaurora.org>