iopoll: remove overly-helpful helper macros, clarify naming
Some of the macros in iopoll.h might be a tad verbose or
redundant. Several of the "verbose" macros are not used anywhere in the
kernel. Based on feedback from upstream, rip out the extra "wrapper"
macros and convert callers (if any) to the lower-level ones.

Also change the `_noirq' suffix to the more idiomatic `_atomic'.

The following semantic patch was used to help identify and make the
necessary changes:

    @@
    expression addr, val, cond, timeout_us;
    @@

    - readl_tight_poll_timeout(addr, val, cond, timeout_us)
    + readl_poll_timeout(addr, val, cond, 0, timeout_us)

    @@
    expression addr, val, cond, max_reads, time_between_us;
    @@

    - readl_poll_timeout_noirq(addr, val, cond, max_reads, time_between_us)
    + readl_poll_timeout_atomic(addr, val, cond, max_reads, time_between_us)

Change-Id: Ibdb054ded59d777f38f594a2f09a12c64abdb059
Signed-off-by: Mitchel Humpherys <mitchelh@codeaurora.org>