usb: gadget: audio: Resolve memory leaks in audio playback and capture
When the USB host initiates USB audio playback of capture, set_alt(1)
command is sent to the device. In response, the USB driver allocates a new
set of USB requests for handling the audio data. This done after each
set_alt(1) command sent from the USB host and the request never get freed.
The fix resolves this issue by ignoring set alt commands if the interface
is already set to this alt interface.

Change-Id: I4ac0acf2ac714bc132cfc0f99c2d04174e31b99b
CRs-fixed: 732934
Signed-off-by: Danny Segal <dsegal@codeaurora.org>