msm: kgsl: Use a lock to protect shared variables in sync timelines
It is possible for concurrent event callbacks to be called on two
timestamps in the same timeline. If the newer timestamp makes it
in first there is a possibility that the older timestamp will
overwrite the last_timestamp variable and cause future comparisons
to fail. Use a spinlock to serialize access to the last_timestamp
variable to make sure that all accesses are clean.

Change-Id: Ic0dedbad339087cc68778c8dbdcbccbc0370d099
Signed-off-by: Jordan Crouse <jcrouse@codeaurora.org>
Signed-off-by: Lynus Vaz <lvaz@codeaurora.org>