regulator: cpr-regulator: fill OPP table with corner to open-loop voltage.
Use linear interpolation of frequencies to calculate the CPR
open-loop voltage for each virtual voltage corner.  Fill an
operating point table associated with the regulator device with
<virtual corner, open loop voltage> pairs.

The open-loop voltage values are used as initial voltages when
CPR is enabled.  They are also used all the time when CPR is
disabled.

Change-Id: I424ee8a6178e85525601a9d24d161e602ac602f9
Signed-off-by: David Collins <collinsd@codeaurora.org>