msm: spm: Configure q2s based on qchannel property when entering PC
On msm8994v1, qchannel should not be ignored when entering power
collapse, due to hardware limitations. Add support for a boolean
property in the spm driver to identify these targets. If this property
is set, the spm driver will configure qchannel to be used when in power
collapse. On targets where this property is not specified, qchannel
will be configured to be ignored when in power collapse.

Change-Id: Ib9e83f7e9a8ed55b88840313e6cc575ea8e791be
Signed-off-by: Karthik Parsha <kparsha@codeaurora.org>