ARM: dts: msm: Update Corner values for Power and CCI for 8939 1.1/2.0
Use below corner values for speed_bin 2 & 4 for better power savings:
	1. Fmax@Power = 17
	2. Fmax@CCI = 17

The Max corner values for each of the client(Perf/power/CCI)
has changed, thus in most of the cases the regulator framework
is unable to find a common max voltage corner to apply and
ends up honoring only the first client. Add dummy Fmax for
Power cluster and CCI to map with max corner of perf cluster
to satisfy the same.

CRs-Fixed: 736378
Change-Id: I04508d103aaacbad048a5fa447a9214819ed89af
Signed-off-by: Tirupathi Reddy <tirupath@codeaurora.org>