power: qpnp-fg: manually clear interrupts at probe time
The fuel gauge does not follow the reboot signal from the msm. This can
cause interrupts to fire in the boot sequence before the kernel can
react.

In particular, this can cause the latched status to be high before the
interrupt is registered with the pending bit low. The SPMI interrupts
driver will be confused and no further interrupts will be serviced from
the fuel gauge.

To avoid this issue, clear the fuel gauge memory interrupts before
registering the interrupts.

CRs-Fixed: 738872
Change-Id: Ic644565e35bfe52627b4905a21ab36f64bd14e71
Signed-off-by: Xiaozhe Shi <xiaozhes@codeaurora.org>