nl80211: check nla_put_* return values
Coverity pointed out that in a few functions we don't
check the return value of the nla_put_*() calls. Most
of these are fairly harmless because the input isn't
very dynamic and controlled by the kernel, but the
pattern is simply wrong, so fix this.

Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Git-commit: ae917c9f55862691e31b84de7ec29bedcb83971c
Git-repo: git://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git
Change-Id: Iefd253fd21f0ceedb915ee2b0de1a22632a30056
Signed-off-by: Ahmad Kholaif <akholaif@codeaurora.org>