ARM: dts: msm: use correct resource size for all IOMMUs
We recently adjusted the size of all IOMMUs on thulium to 0x20000
[78412efa77a: "ARM: dts: msm: increase resource size of IOMMUs"] since
that's the size given by some hardware documentation.  However, the size
of each SMMU is actually a function of how many implemented context
banks it has, so increasing the size of *all* the SMMUs actually fixed
some of them and broke others.

Fix this by using the correct resource size for all SMMUs (0x10000 or
0x20000) depending on how many context banks they have.

Change-Id: I64d5c2af8f713e690a4316908d0d704c21aa6085
Signed-off-by: Mitchel Humpherys <mitchelh@codeaurora.org>