power: qpnp-fg: do not reload identical profiles
Currently the fuel gauge driver always reloads the battery profile upon
probe. Reloading the battery profile will also necessitate an SOC
reestimation. This is undesirable because a settled SOC is usually more
accurate than a first time estimation.

Fix this by comparing the battery profile loaded in SRAM with the one
picked by the fuel gauge driver. If they are identical, then do not
force an SOC reestimation and do not reload the profile.

CRs-Fixed: 720242
Change-Id: I50662d5e5c14b9513fbfcd06b964615db6fcde7e
Signed-off-by: Xiaozhe Shi <xiaozhes@codeaurora.org>
[imaund@codeaurora.org: Resolved conflicts related to renaming of
  INIT_COMPLETION to reinit_completion.]
Signed-off-by: Ian Maund <imaund@codeaurora.org>