iommu: msm: actually disable L2 redirect when CONFIG_IOMMU_PGTABLES_L2=n
We recently moved the L2 redirect configuration to a domain attribute
[40cc2c777c54b: "iommu: msm: move L2 redirect to a domain attribute"].
This allows us to do away with some divergence in the generic IOMMU
APIs.  However, that commit botched the case when
CONFIG_IOMMU_PGTABLES_L2 is disabled.  In that case we actually always
set the page tables to be shared, exactly the opposite of what we want.
Fix this.

Change-Id: I47837584ae88fbc0be578500d20c2a62a9b33bca
Signed-off-by: Mitchel Humpherys <mitchelh@codeaurora.org>