msm: kgsl: Hold the process mutex when destroying the sysfs nodes
Each process may have several threads opening and closing the kgsl
device. Hold the process mutex while destroying the process's sysfs
nodes to keep the them in sync with the process_priv's status in the
list.

This change also removes the process_priv from the process list when
the refcount becomes zero. The kgsl device may be opened from
multiple threads in a process, so the process_priv needs to be
removed from the list only when the last thread closes it.

Change-Id: Ibad15c7b1c272b267e3df06112ea487e1988a28e
Signed-off-by: Lynus Vaz <lvaz@codeaurora.org>