usb: dwc3-msm: Cancel block reset while cable disconnect
There might be a race between block reset due to an ERRATIC
event and cable disconnect at the same time. This might lead to an
unclocked register access in dwc_gadget_run_stop().
To avoid this race, whenever there is a cable disconnect while block reset
in progress, cancel block reset work and wait for it to finish.

CRs-fixed: 739944
Change-Id: Id0e0fc7faae82b7c52360c6247a6eaa92d4f9997
Signed-off-by: Azhar Shaikh <azhars@codeaurora.org>