ARM: dts: msm: add mdss smmu context banks mapping for thulium target
Thulium target uses smmu v2 in which each context bank is handled
as a separate device. These devices are added as part of the mdss_mdp.
It also requires clk voting to be handled before accessing the smmu
region, so each context bank has its own clks which is associated
with the device.

Change-Id: I273af1800543395f5b144aa9fa5b6483468992b6
Signed-off-by: Veera Sundaram Sankaran <veeras@codeaurora.org>