sched: adjust small task wakeup preference, add sync wakeup support
Small tasks currently wake up an idle CPU if a mostly idle CPU (that
is not in low power mode) is not found in the lowest power band.
This may cause costly latency penalties.

Instead, search the whole CPU topology for a mostly idle CPU which
is not in a low power state and choose one of these prior to
resorting to other options.

Also, calculations as to whether a CPU is mostly idle needs to take
into account whether the wakeup is a sync wakeup. If it is a sync
wakeup, then the currently running task on the CPU needs to be
discounted from the CPU's load as we check on whether it is
mostly idle.

Change-Id: Ibf293e1c4a7ec0c4e091200d95e97f08cbe4116c
Signed-off-by: Steve Muckle <smuckle@codeaurora.org>