sched: add sync wakeup recognition in select_best_cpu.
If a wakeup is a sync wakeup, we need to discount the currently
running task's load from the waker's CPU as we calculate the best
CPU for the waking task to land on.

Also fix an oversight in the small task wakeup path, the spill
threshold check there was not sync-wakeup aware.

Change-Id: I00c5df626d17868323d60fb90b4513c0dd314825
Signed-off-by: Naveen Kumar <kumarn@codeaurora.org>