regulator: cpr-regulator: add support for initial voltage adjustment.
Add support for initial voltage adjustment that is decided based
upon the values read out of a combination of fuse parameters.
This allows for precise control of initial voltages in order to
compensate for differences in characterization procedures used
along with specific CPR fuse revision and ring oscillator select
values.

Change-Id: I108931f2abe4186307c397e229e0fe34bf2761c6
Signed-off-by: David Collins <collinsd@codeaurora.org>