msm: rndis: move resource manager setup to connect/disconnect events
This change moves the IPA Resource Manager setup API to
connect and disconnect events.
it is done in order to have the same behavior as of ECM_IPA.
The same behavior should apply to both ECM_IPA and RNDIS_IPA due
to the fact that they use the same API of IPA Resource Manager
which has internal dependencies of these APIs timing, thus, the ECM setup
and RNDIS setup should happen at the same sequence.

Change-Id: Ie6cea28ced06194f747b3773d5f119664794e83f
Signed-off-by: Talel Shenhar <tatias@codeaurora.org>