ARM: dts: msm: Change BIMC BWMON interrupts to level triggered.
Since the BIMC BWMON IRQ is a shared interrupt across the ports, it needs
to be treated as a level triggered IRQ to avoid races.
Otherwise, the following race can happen:
* IRQ arrives
* Port 0 checks if it caused the IRQ.
* Decides it did not and doesn't do anything.
* Port 1 checks if it caused the IRQ.
* Decides it did cause it and is processing it.
* Port 0 now triggers the IRQ.
* Port 1 handler clear its IRQ source.
* IRQ line remains high throughout this period.
* All IRQ handlers are done processing the rising edge triggered IRQ and
no more work is done.
If this is instead handled as a level triggered IRQ, then the handlers
will be called again and the Port 0 handler would process its IRQ and then
clear its IRQ source and the IRQ line would go low.

Change-Id: Ic4a2f2fbf5348dfa6f637ca4ec317da1af728358
Signed-off-by: Panwar Vivek <vpanwa@codeaurora.org>