iommu: make iommu_unmap_range return 0 on success
The .unmap_range callbacks in the various MSM IOMMU driver
implementations have historically returned 0 on success, and non-zero on
failure.  However, the generic implementation of unmap_range returns the
number of bytes unmapped.  Since we're soon going to be getting rid of
iommu_unmap_range anyways just make the generic unmap_range
implementation return 0 on success and non-zero on failure to emulate
the old way of doing things.  That way clients only have to change their
code once (which will be when we delete iommu_unmap_range).

Change-Id: Iedc524b248197de59bdbb21069be773e4b2fe68a
Signed-off-by: Mitchel Humpherys <mitchelh@codeaurora.org>