ASoC: msm8x16-wcd: acquire lock in irq_handler
During suspend, if button is pressed button press handler is
waiting for codec resume to handle interrupt, meanwhile if button is
released, release handler is proceeding further as codec is resumed.
Due to this driver is reporting release event an then press event.
Avoid this by acquiring lock in irq handler.

CRs-Fixed: 754472
Change-Id: Iec398863583fc771dde65bf66f239edfa974b743
Signed-off-by: Yeleswarapu Nagaradhesh <nagaradh@codeaurora.org>