ARM: dts: msm: Set minimum CPUBW based on CPU freq for MSM8939
This allows setting a DDR frequency floor based on CPU frequency is
addition to using the bw_hwmon governor for decide the final CPU BW vote.
Because of how the BW votes are aggregated, if bw_hwmon pick a lower vote
than the floor, the final vote will be the floor value.

Change-Id: I5560c573818b386264025948c9162d816e19798b
Signed-off-by: Taniya Das <tdas@codeaurora.org>