iommu/arm-smmu: enforce valid usage of coherent table walk domain attr
Currently we expose a domain attribute to disable/enable coherent
hardware table walks.  However, while we're building page tables, we
flush the cache based on the SMMU's capability to perform coherent
hardware table walks, rather than flushing based on whether coherent
hardware table walks are actually enabled for the domain.  We also allow
domains that have asked to enable coherent hardware table walks to be
attached to SMMU devices that don't support it.

Fix these issues by doing validation of the domain attributes against
the SMMU device to which the domain is attached.  We need to do this at
domain attribute set time and at attach time since domain attributes can
be set before and after they are attached to a device.

This also allows us to drop the check for whether the SMMU supports
coherent table walks in the page table flushing code.  We can simply
check the domain attributes instead of checking both the domain
attributes and the SMMU features.

CRs-Fixed: 752812
Change-Id: Ibee9316a340f481664f5fb7757a24f136e0ed74c
Signed-off-by: Mitchel Humpherys <mitchelh@codeaurora.org>