soc: qcom: glink_smd_xprt: Handle same named channel on multiple edges
SMD driver registers a platform device to indicate the availability of a
channel. SMD channels are unique by name/edge tuple, but the platform
driver can only specify the name of the channel. Add logic to manage the
platform drivers, so that multiple SMD channels with the same name towards
different edges can be handled appropriately.

Change-Id: I6723c6def832200196bd431737e3f10ffac30093
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>
Signed-off-by: Karthikeyan Ramasubramanian <kramasub@codeaurora.org>