ARM: dts: msm: correct the apss funnel inport info
There is a mismatch in the number of inports specified for apss funnel for
both 8994v1 and v2. This apss funnel is used in ETM tracing path and
incorrect funnel inport info could cause data corruption and issues with
ETM trace for cores 4 to 7. Correct the apss funnel inport info for apss
funnel in device tree.

Change-Id: Ib7503ca9dd34a8fd4a5b396baef91a775f0c2ef7
Signed-off-by: Sarangdhar Joshi <spjoshi@codeaurora.org>