msm-core: Add uio extension for memory mapping
With /dev/mem node no longer available for userspace IO memory access,
each driver needs to provide the range of memory it needs access to and
the type of memory it is accessing.

Add this interface to the driver and create a device that allows
userspace to access IO memory mapped region. Since userspace access is
simple IO memory read, use the generic UIO ops function for mmap and
open calls.

Change-Id: I34ffd2873ed7998244bded7f88f6663dd95c6ad4
Signed-off-by: Archana Sathyakumar <asathyak@codeaurora.org>
[imaund@codeaurora.org: Resolved context conflicts]
Signed-off-by: Ian Maund <imaund@codeaurora.org>