regulator: cpr-regulator: Add support for fuse based quotient offset
In some SOCs, quotient offsets per fuse corner can be programmed
in the fuse so that the software can make use of them without the
need to calculate them. Add support to read those quotient offsets
which are defined per fuse corner and use them to obtain the
quotient scaling factors, so that the quotient for a particular
frequency can be calculated.

In addition, add support to adjust target quotient offset based
on the values read from device tree.

Change-Id: I63c3376ea43eaff276bdeb0fdd7000c024abd46e
Signed-off-by: Subbaraman Narayanamurthy <subbaram@codeaurora.org>
[imaund@codeaurora.org: resolved context conflicts]
Signed-off-by: Ian Maund <imaund@codeaurora.org>