ARM: dts: qcom: Correct regulator settings for SSC subsystem node
Do not try to set voltage of the LVS2 regulator as it is really
just a switch which can only be enabled or disabled. The output
voltage will be fixed at 1.8V.
In addition, list the regulators for the SSC node on Thulium
under active-reg-names instead of proxy-reg-names as they need
to be turned on when the subsystem is online.

Change-Id: Iac40264a5c90ba3303c66b42ea73211359cbef43
Signed-off-by: Deepak Katragadda <dkatraga@codeaurora.org>