msm: kgsl: Correct payload size for type 4 and type 7 CP packets
The new PM4 type4 and type7 packets now use the actual size of the
payload while the legacy types use the size of the payload minus 1.
Unfortunately the newly added cp_type4_packet and cp_type7_packet
macros still subtracted 1 from the count which forcing the developer
to remember to add 1 to the count when calling the macro.

Fix the macros so the passed in size for all PM4 macros is the size
of the payload minus the header.

Change-Id: Ic0dedbad91d9dbdbd762d89993ffcae671a91c1a
Signed-off-by: Jordan Crouse <jcrouse@codeaurora.org>