power: qcom: Add Array Power Mux driver
Introduce a driver to interface with the APM controller which
is responsible for selecting the power rail sourcing the CPU and
cache memory arrays. Support an API that performs the programming
steps to switch the rail powering these arrays in order to fulfill
the SRAM Vmin requirement across different performance levels.

Change-Id: I7659c00b68c5cc61653c84cb14a30a8ef3c29cca
Signed-off-by: Osvaldo Banuelos <osvaldob@codeaurora.org>