defconfig: msm: enable MSM APM driver
The MSM APM driver is used on some MSM platforms to control
the voltage source powering the cpu and cache memory arrays. Since
the SRAM arrays have a minimum operational voltage, switching
between the rails powering these is necessary as the rails
can be scaled scaled up or down and cross the Vmin threshold.

Change-Id: Ia0d3eb654cdafb81eb3ad4aa0a5e2e61c4587add
Signed-off-by: Osvaldo Banuelos <osvaldob@codeaurora.org>