clk: clock-local2: Allow not checking status bit in the branch disable op
When disabling a branch clock, if the no_halt_check_on_disable
flag is set, do not check for the status bit to change.
With the new SMMU vote register support on Thulium, although
the HLOS clock driver might try and disable the clock, it might
not turn off because of over-laying votes from TrustZone or
Hypervisor. Hence, do not check for the status bit to change
while disabling these clocks.

Change-Id: I86316bf5519660bd4ca871fd832f4598ed6b9b1a
Signed-off-by: Deepak Katragadda <dkatraga@codeaurora.org>