usb: dwc3: core: Define DTCT field in GUCTL register
Define Device Timeout Coarse Tuning field in Global
User Control register and document the need for
keeping value at POR ie. 0.

The default value of 0 for the DTCT is sufficient to
meet the timing requirements specified in the USB 3.0 specification.

If a non-zero setting is used with the xHCI driver, the device
could be re-enumerated.

CRs-fixed: 754911
Change-Id: I7f53ff23066649bd4b84bfc7177d07ff495bd2d8
Signed-off-by: Devdutt Patnaik <dpatnaik@codeaurora.org>
Signed-off-by: Mayank Rana <mrana@codeaurora.org>