msm: ipa: destroy QMI client handle in IPA QMI clean-up
This change fixes a QMI client handle leak found in IPA.
The issue is caused due to the fact that during the clean-up,
rmnet_ipa unregisters from QMI events, causing QMI_SERVER_EXIT
to not be called, and handle clean-up to not be invoked.
The change invokes the handle clean-up explicitly.

CRs-fixed: 767260
Change-Id: I4c2047d08d9c9608ab80dd2673983794832f9ad5
Acked-by: David Arinzon <darinzon@qti.qualcomm.com>
Signed-off-by: Sivan Reinsten <sivanr@codeaurora.org>