soc: qcom: msm_performance: Maintain correct CPU offline status
As per the current code, if any other entity decides to offline a
CPU and the module disallows it from coming up to meet the current
request, when cluster control is released the CPU in question might
be left offline since the module only plugs back in cores which it
itself offlined. Hence it is crucial to maintain a correct status of
offlined cores to be brought back up online on control release request.

This change identifies the CPU that module prevents from being onlined
so that it is restored back when cluster control is released.

Change-Id: I87831eca879d1cf7d33f88a881e0427641e39bd6
Signed-off-by: Rohit Gupta <rohgup@codeaurora.org>
Signed-off-by: Niranjan Reddy Dumbala <ndumba@codeaurora.org>