ARM: dts: msm: Fix the igc offset for source pipes in thulium
IGC(inverse gamma correction) block offset is stored in the device tree
since offsets can change based on the mdp version. For thulium the IGC
offset in device tree was wrong which was leading to incorrect programming
of IGC block. This change fixes the IGC offset in the device tree.

Change-Id: Ic8a65ac3016bfbc27d647bbe5dff601d0d208a6d
Signed-off-by: Gopikrishnaiah Anandan <agopik@codeaurora.org>