thermal: tsens: Update interrupt handling to threaded irq
Update registering the TSENS IRQ with request_threaded_irq()
instead of request_irq() to execute checking for the sensors
threshold. This is done to run with SCHED_FIFO and MAX_RT_PRIO/2
by default which request_threaded_irq() provides on the
threaded function to improve the response to notify clients
for temperature threshold crossing. Use the IRQF_ONESHOT flag
while registering the interrupt to keep the irq line at the
interrupt controller disabled until the threaded handler has been run.

Change-Id: I9dec7a21719a0a6a6f0bcbac3221b4a4bfdca322
Signed-off-by: Siddartha Mohanadoss <smohanad@codeaurora.org>
[imaund@codeaurora.org: Resolved context conflicts and conflicts caused
  by the new use of devm_request_irq in 3.14.]
Signed-off-by: Ian Maund <imaund@codeaurora.org>