usb: msm: bam: Handle multiple bam pipe suspend/resume properly
Driver maintains pipes_suspended and pipes_resumed counter to
match the bam pipe suspend and resume. Bam pipes are suspended
one by one but condition to check if all the bam pipes are
suspend and resumed is incorrectly assuming that pipes are
suspended in pairs. As a result pipe_suspended and pipes_resumed
values are set less than the actual number of pipes suspended
and resumed. At the time of resume all suspended pipes are not
reset because wrong value of pipe_suspended is set. This is causing
data stall. Hence remove multiplication of 2 while checking for
number of pipes suspended and resumed.

CRs-Fixed: 773747
Change-Id: I9746e9c9f7596c5e4737c50a07dc63dd60738ba6
Signed-off-by: Hemant Kumar <hemantk@codeaurora.org>