regulator: cpr-regulator: Add minimum difference allowed for quotients
Sometimes after power-optimization adjustments to the fused quotients
the values may end up in a non-increasing order (NOM > TURBO) or
may get very close to each other. In the former case the driver
disables CPR. To avoid this, add support for specifying a minimum quotient
difference between 2 fuse corners. If this difference is not met add a
configurable quotient offset to the fuse corner quotient.

Example:
	cpr-fuse-min-quot-diff = <0 0 40>;
	cpr-min-quot-diff-adjustment = <0 0 104>;

	CPR log snippet:
		Corner[2] adjusted quotient = 1275
		Corner[3] adjusted quotient = 1261

	After adjustment applied:
		Corner[3] revised adjusted quotient = 1379

Change-Id: Iec8b9a703a0018620f6d1d23326e6a32c3bf88d0
Signed-off-by: Tirupathi Reddy <tirupath@codeaurora.org>