Revert "USB: dwc3: otg: Turn off VBUS before removing hcd"
This reverts commit 0d3c96aa1cf53596f73845b3247988d4e9743bff.

The original snapshot of dwc3_otg.c already contained this line
of code, and this patch was inadvertently cherry-picked, causing
a duplicate call to platform_device_del(). This leads to a NULL
pointer dereference during the sysfs cleanup path during the
second call, and results in a panic whenever a role switch occurs
from USB host mode back to device mode.

Change-Id: I1ec5c995e76235e755f34b43fb84c0682cda4315
Signed-off-by: Jack Pham <jackp@codeaurora.org>