usb: dwc3: get rid of otg-capability
Remove the usage of the 'otg-capability' flag. Its usage has
evolved to become an indicator that dual-role support is
present, along with external VBUS detection for peripheral mode.
Furthermore, when it is false, it uses the untested and likely
broken OTG support built-in to the hardware, which somewhat
contradicts the name of this flag. Since this flag is always true
for current targets, remove it along with the unused OTG IRQ and
registers to allow for code simplification.

Change-Id: Ie8a0e9e60ef53e29f49966cf54b5a8ef1bb68be8
Signed-off-by: Jack Pham <jackp@codeaurora.org>