clk: msm: clock-local2: Account for non-locally-controlled RCG branches
Some RCGs may have child branches; at least one of which is controlled
by an execution entity other than the apps processor. This requires
some special handling; ensure that mux writes are staggered with the
enable/disable signals, and that when _this_ execution entity disables
the RCG - i. e., removes voltage votes and disables sources - that the
RCG mux is set to a safe source.

CRs-Fixed: 690647
Change-Id: Ie6be17981ba86fe312076a1e091a215198e2c153
Signed-off-by: Vikram Mulukutla <markivx@codeaurora.org>