ARM: dts: msm: change system power management level to default
UFS controller on MSM Thulium supports the UFS power collapse feature
when link is in hibern8 so there is no need of putting the UFS link
in OFF state during system suspend. With this change, UFS system PM
level will be set to default which allows hibern8 with power collapse.

Change-Id: Iaef77f95485d1b7dbb944d5792c9668341fe4c10
Signed-off-by: Subhash Jadavani <subhashj@codeaurora.org>