ASoC: wcd_cpe_services: Change boot event to include SFR details
In case of CPE crash, it is possible to retrieve the SFR (Software
Failure Reason) details from CPE. Add the support in the CPE boot event
to parse the SFR buffer address and size after CPE bootup is done. In
case of crash, the SFR data will be read from this address.

Change-Id: I8802e844f10d15fd14080cc7597ebfde0e5073d7
Signed-off-by: Bhalchandra Gajare <gajare@codeaurora.org>