soc: qcom: glink: Channel migration corner case fixes
If the local side is migrating to another transport, and the remote side
has created a context on that transport first, a duplicate context will be
created and the channel will not connect after migration.  Also if a new
transport comes up after the local side has opened the channel, and the
remote side wants to use the new transport, the local side won't migrate
properly because its vote is based on stale information.

Fix these issues by merging the remote context into the local context when
duplication will occur in migration, and updating the local vote if
necessary.

CRs-Fixed: 786826
Change-Id: I8adcb0c57130b7f062dc48d7df622a27306cea97
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>