soc: power: pm: Add documentation for sleep status check
During hotplug, drivers wait for core to be power collapsed to turn off
the resources used by a core in CPU_DEAD notification. The kernel
thermal module which hotplugs the core in certain conditions is
registered in the arch_initcall(). To ensure that the CPU_DEAD waits for
sleep status to be set during hotplug, register the cpu sleep status
driver before the regulator driver.

To this effect, the sleep status driver is no longer made a child device
for pm-8x60 device.

Change-Id: Ie8d23f120879dc877e5a2e4ecf4915ad112b17a4
Signed-off-by: Archana Sathyakumar <asathyak@codeaurora.org>