msm: kgsl: The CP_PROTECT_XX registers on A430 are not contiguous
The A430 chipset added a new block of CP_PROTECT_ registers but
because of the existing map the new registers are not contiguous.
Unfortunately the code assumed that they were so the register
in the middle (CP_PROTECT_CTRL) was getting overwritten.

Adjust the offset for the upper bank of CP_PROTECT_ registers
on A4XX targets.

Change-Id: Ic0dedbad871395467a2cf3d3b0115f13d60256ce
Signed-off-by: Jordan Crouse <jcrouse@codeaurora.org>