power: qpnp-smbcharger: fix the argument order while using is_between()
The is_between() function expects the value to be the third argument, the
first and the second arguments define the range.

Currently the value to be checked is passed as the first argument which
leads to selecting wrong voltage thresholds for Wipower dynamic
impedance limiting feature which in turn causes an ADC interrupt storm.

Fix this by passing arguments in the right order.

Change-Id: I525615ca2a2f4a4e3a0db15eb3e15ec965853552
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>
Signed-off-by: Xiaozhe Shi <xiaozhes@codeaurora.org>