ARM: dts: msm: add pinctrl configs for HDMI on thulium CDP and MTP
Add HDMI pinctrl configs in thulium which will be used for CDP and MTP.
Four pinctrl states are defined which will be used based on mode of
operation.

hdmi_active state is enabled when HDMI cable is connected and CEC is
enabled. hdmi_sleep state is used when hot plug detection is not required
and the controller is disabled. hdmi_ddc_active state enables HPD and DDC
lines. hdmi_cec_active enables HPD and CEC lines.

Change-Id: Iddc3dc6deb73d9ec1206a23bcbc809158aad43dc
Signed-off-by: Vinu Deokaran <vinud@codeaurora.org>