USB: gadget: Introduce a function workqueue in usb_gadget
In very fast back to back suspend resume scenario, we are encountering
a case where a sequence of supend resume suspend happens. While first
suspend work is still running we get resume and second suspend as well.
As second resume has happened we see that A_BUS_REQ is clear, when trying
to enter LPM and enter LPM. After that resume work is queued and due to
unclocked access causing target to crash.

Avoid the above mentioned case by introducing a function workqueue, which
will be used to queue these work. This workqueue will be flushed as part
of msm_otg suspend before going to LPM.

Change-Id: Ie364d8c2ee1bfe7adb790bdc8717b833cd2f5d97
Signed-off-by: Tarun Gupta <tarung@codeaurora.org>