thermal: tsens: Interrupt updates for thulium
Fix the mask used to set the corresponding sensors
for Upper/Lower/Critical interrupt that crossed the
temperature threshold. Update the Upper/Lower and
Critical interrupt to register it as a threaded_irq.
This is done to run with SCHED_FIFO and MAX_RT_PRIO/2
by default which request_threaded_irq() provides on
the threaded function to improve response to notify
the clients for temperature crossing.

Change-Id: If9b537ef9b5343abbeb871b3c9b69b362f15ab11
Signed-off-by: Siddartha Mohanadoss <smohanad@codeaurora.org>