clk: msm: gdsc: Dont check the status bit when disabling votable GDSCs
Do not poll the status bit when the GDSC is disabled. With the
new SMMU vote register support on Thulium, although the HLOS
gdsc driver might try and power collapse the GDSC's power domain,
it might not collapse because of overriding votes from secure
world or hypervisor. Hence, do not perform the status bit polling
when disabling these gdscs; instead, add a short delay here to
ensure that if we were truly the last ones to disable the GDSC,
the state machine has enough time before we enable it again.

Change-Id: I434b11d837d0c7e2f6eeadfe76e0758acc8eebc8
Signed-off-by: Deepak Katragadda <dkatraga@codeaurora.org>