ARM: dts: msm: Add OTG related regulators for msmthulium
During DC charging, some devices may want to support USB OTG
with an external boost. The PMI8994 has a 5V audio boost that
can be utilized  to provide the 5V that OTG requires.

Add the usb_otg_switch fixed regulator to enable the switch that
connects the boost with USBIN via PMI8994 GPIO 5.

Add the smbcharger_external_otg regulator to disable the usb
charge path when the 5V is provided.

The layout of the regulator supply chain will be the following:
	- usb_otg_switch
		- fixed regulator that controls gpio to enable or
		disable OVP FET
	- supplied by: smbcharger_external_otg
		- smbcharger regulator that disables usb charging
		and USB source detect
	- supplied by: pmi8994_boost_5v
		- rpm regulator that enables the 5V smart boost in
		the PMI8994.

The usb_otg_switch can be used in place of smbcharger_charger_otg
to provide the 5V in place of the charger operating in reverse if
specified in the device tree.

Change-Id: Ib712652deaecded32002dd6f3b51438bf74a441a
Signed-off-by: Osvaldo Banuelos <osvaldob@codeaurora.org>