soc: qcom: cache_m4m_erp64: log messages when interrupts are registered
If errors are detected by hardware prior to the driver probe, they
will be reported immediately following registration for the error
interrupt.

Log informational messages at probe time just before registering
for these interrupts, to serve as a clue that the error may have
occurred quite a while before it was actually reported.

Change-Id: I59e0623e1120fd507d1045b0288ea4d498e600eb
Signed-off-by: Matt Wagantall <mattw@codeaurora.org>