msm: kgsl: Disable L2 bypass on A5XX since it is broken
The UCHE_TRAP_BASE register wasn't being correctly programmed.
The area between the UCHE_WRITE_THRU_BASE and the
UCHE_TRAP_BASE defines the L2 bypass region which is broken on
A5XX. As a result we program the two to be equal to each other
as a result disabling L2 bypass. In addition, enable the UCHE
TRAP interrupt to catch such a problem sooner in the future.

Change-Id: I31f0982d5d4c3ee51052c42ad0f99f5bb7bc2543
Signed-off-by: Shrenuj Bansal <shrenujb@codeaurora.org>